feat(logbook): migrate state to Pinia
