--- src/store/logbook/state/index.js
@@ -1,11 +1,0 @@
-export default {
-  grades: {},
-  styles: {},
-  crags: {},
-  sectors: {},
-  routes: {},
-  ascents: {},
-  data: {},
-
-  dataRetrieved: false,
-};
