--- src/store/logbook/types.js
@@ -1,15 +1,0 @@
-export const Action = {
-  GET_DATA: "GET_DATA",
-};
-
-export const Mutation = {
-  SET_ASCENTS: "SET_ASCENTS",
-  SET_CRAGS: "SET_CRAGS",
-  SET_DATA: "SET_DATA",
-  SET_DATA_MISSING: "SET_DATA_MISSING",
-  SET_DATA_RETRIEVED: "SET_DATA_RETRIEVED",
-  SET_GRADES: "SET_GRADES",
-  SET_ROUTES: "SET_ROUTES",
-  SET_SECTORS: "SET_SECTORS",
-  SET_STYLES: "SET_STYLES",
-};
